Move ExtractCityFromMessagePrompt to MessageHelpers
